🐛 fix: suspense
